The index of playerContainer now sorts correctly. Removed std:: from the program.
Instead of using std::ITEM everywhere, header files now include the directive
"using std::ITEM" for each ITEM they need. This cleans up the code,
and is also better form than writing "using namespace std".

The playerContainer class now properly handles the index
member. It is assigned when the default or player constructors are called,
and is equal to a static count of how many instantiated playerContainer
objects exist. When the copy constructor or assignment operator are called,
the index member is copied as well.