Added NeuralPlayer and ManualPlayer derived classes.
The NeuralPlayer class is controlled by its internal neural network. This is tuned by the genetic algorithm. This class has a public neuralNetwork member which holds the internal weights. However, a workaround must be used to access this throught the rest of the program.

    This is because the playerContainer objects have a Player* member. Although this pointer setup is useful for allowing any derived class to be stored in it, it can't call or access any derived class members... even if it is an instance of that derived class. This is unless it is explicitly cast to that type's pointer, which is what I did for now. I did it with dynamic_cast<T*> and then checked T != NULL, or else I exit the program.

The ManualPlayer class is controlled by user input either through the console or a file. It can be used to quickly judge the basic sensibility of the popultaion (i.e. is it terrible, decent, or excellent). It has ostream& and istream& members for i/o, and reads from the istream to make its next move.