playerContainer objects are now templated.
playerContainer objects have a member variable which holds the particular type of player that object has. Before, that was handled by a pointer to the abstract class Player. This worked because all derived classes are still of type base class. However, with the new templated implementation the type of the member variable is determined at compile time. This gets rid of needing to perform static casts to the correct type of player during runtime.