Finished integrating the templated Player class. Implemented a player Hall of Fame.
The Hall of Fame stores the best player from each generation. This can be used as a benchmark to track long term progress of the population.

I also began work on a PerfectPlayer class (as another way of tracking progress), although I'm not sure if I will continue working on it.

I also deleted TicTacToe.cpp. This is because I made the TicaTacToe class templated, and templated classes need all there code in the same file. Now only the TicTacToe.h file exists (it can still be included like before though).