Main program handles loading and saving players.
The main program asks whether the user wants to load a saved player or continue as normal. If they load they type in the name of the log file to load. If not, the training progresses as usual. At the end, the user enters the name of the log file to be saved.

Added .log types to .gitignore